requirements.txt
agno>=1.2.6
python-telegram-bot>=22.0
google-auth>=2.38.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.166.0
python-dotenv>=1.1.0
pydantic>=2.11.1
httpx>=0.28.1
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.32.3
pytest>=7.4.0 